Tried to fix the warning from github for actionview.
I performed bundler update actionview but this did not change. Instead nokogiri was bumped up to 1.10.9 from 1.10.8. I will wait and see when merging with master to deal with the warning.

Dependabot cannot create a pull request as one or more other dependencies require a version that is incompatible with this update.